+++ src/theme-toggle.js
@@ -5,9 +5,7 @@ const template = document.createElement('template');
 template.innerHTML = /* html */`
   <style>
     :host {
-      --icon-size: 24px;
-      --icon-color: currentColor;
-
+      display: inline-block;
       box-sizing: border-box;
     }
 
@@ -25,15 +23,19 @@ template.innerHTML = /* html */`
       display: inline-flex;
       justify-content: center;
       align-items: center;
-      gap: 0.25rem;
-      padding: 0.5rem;
+      gap: 0.375rem;
       border: none;
-      border-radius: 0.25rem;
       background-color: transparent;
       color: inherit;
       cursor: pointer;
       font-family: inherit;
       font-size: inherit;
+    }
+
+    .default-icon {
+      min-width: 1.5em;
+      width: 1.5em;
+      height: 1.5em;
     }
 
     .system {
@@ -42,50 +44,47 @@ template.innerHTML = /* html */`
     }
   </style>
 
-  <button type="button" part="button" id="theme-toggle" aria-live="polite">
+  <button type="button" part="base" id="theme-toggle">
     <slot name="light">
-      <svg part="icon-light" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" width="24" height="24" fill="var(--icon-color)" style="width: var(--icon-size); height: var(--icon-size);">
-        <path d="M12 7c-2.76 0-5 2.24-5 5s2.24 5 5 5 5-2.24 5-5-2.24-5-5-5zM2 13h2c.55 0 1-.45 1-1s-.45-1-1-1H2c-.55 0-1 .45-1 1s.45 1 1 1zm18 0h2c.55 0 1-.45 1-1s-.45-1-1-1h-2c-.55 0-1 .45-1 1s.45 1 1 1zM11 2v2c0 .55.45 1 1 1s1-.45 1-1V2c0-.55-.45-1-1-1s-1 .45-1 1zm0 18v2c0 .55.45 1 1 1s1-.45 1-1v-2c0-.55-.45-1-1-1s-1 .45-1 1zM5.99 4.58c-.39-.39-1.03-.39-1.41 0-.39.39-.39 1.03 0 1.41l1.06 1.06c.39.39 1.03.39 1.41 0s.39-1.03 0-1.41L5.99 4.58zm12.37 12.37c-.39-.39-1.03-.39-1.41 0-.39.39-.39 1.03 0 1.41l1.06 1.06c.39.39 1.03.39 1.41 0 .39-.39.39-1.03 0-1.41l-1.06-1.06zm1.06-10.96c.39-.39.39-1.03 0-1.41-.39-.39-1.03-.39-1.41 0l-1.06 1.06c-.39.39-.39 1.03 0 1.41s1.03.39 1.41 0l1.06-1.06zM7.05 18.36c.39-.39.39-1.03 0-1.41-.39-.39-1.03-.39-1.41 0l-1.06 1.06c-.39.39-.39 1.03 0 1.41s1.03.39 1.41 0l1.06-1.06z"></path>
-      </svg>
+      <slot name="icon-light" part="icon icon-light">
+        <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" width="24" height="24" fill="currentColor" class="default-icon">
+          <path d="M12 7c-2.76 0-5 2.24-5 5s2.24 5 5 5 5-2.24 5-5-2.24-5-5-5zM2 13h2c.55 0 1-.45 1-1s-.45-1-1-1H2c-.55 0-1 .45-1 1s.45 1 1 1zm18 0h2c.55 0 1-.45 1-1s-.45-1-1-1h-2c-.55 0-1 .45-1 1s.45 1 1 1zM11 2v2c0 .55.45 1 1 1s1-.45 1-1V2c0-.55-.45-1-1-1s-1 .45-1 1zm0 18v2c0 .55.45 1 1 1s1-.45 1-1v-2c0-.55-.45-1-1-1s-1 .45-1 1zM5.99 4.58c-.39-.39-1.03-.39-1.41 0-.39.39-.39 1.03 0 1.41l1.06 1.06c.39.39 1.03.39 1.41 0s.39-1.03 0-1.41L5.99 4.58zm12.37 12.37c-.39-.39-1.03-.39-1.41 0-.39.39-.39 1.03 0 1.41l1.06 1.06c.39.39 1.03.39 1.41 0 .39-.39.39-1.03 0-1.41l-1.06-1.06zm1.06-10.96c.39-.39.39-1.03 0-1.41-.39-.39-1.03-.39-1.41 0l-1.06 1.06c-.39.39-.39 1.03 0 1.41s1.03.39 1.41 0l1.06-1.06zM7.05 18.36c.39-.39.39-1.03 0-1.41-.39-.39-1.03-.39-1.41 0l-1.06 1.06c-.39.39-.39 1.03 0 1.41s1.03.39 1.41 0l1.06-1.06z"></path>
+        </svg>
+      </slot>
+      <slot name="label-light" part="label label-light">Light theme</slot>
     </slot>
 
     <slot name="dark">
-      <svg part="icon-dark" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" width="24" height="24" fill="var(--icon-color)" style="width: var(--icon-size); height: var(--icon-size);">
-        <path d="M9.5 2c-1.82 0-3.53.5-5 1.35 2.99 1.73 5 4.95 5 8.65s-2.01 6.92-5 8.65c1.47.85 3.18 1.35 5 1.35 5.52 0 10-4.48 10-10S15.02 2 9.5 2z"></path>
-      </svg>
+      <slot name="icon-dark" part="icon icon-dark">
+        <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" width="24" height="24" fill="currentColor" class="default-icon">
+          <path d="M9.5 2c-1.82 0-3.53.5-5 1.35 2.99 1.73 5 4.95 5 8.65s-2.01 6.92-5 8.65c1.47.85 3.18 1.35 5 1.35 5.52 0 10-4.48 10-10S15.02 2 9.5 2z"></path>
+        </svg>
+      </slot>
+      <slot name="label-dark" part="label label-dark">Dark theme</slot>
     </slot>
 
     <slot name="system">
-      <svg part="icon-system" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" width="24" height="24" fill="var(--icon-color)" style="width: var(--icon-size); height: var(--icon-size);">
-        <path d="M20 3H4c-1.1 0-2 .9-2 2v11c0 1.1.9 2 2 2h3l-1 1v2h12v-2l-1-1h3c1.1 0 2-.9 2-2V5c0-1.1-.9-2-2-2zm0 13H4V5h16v11z"></path>
-        <path class="system system--light" d="M12 7c-2.76 0-5 2.24-5 5s2.24 5 5 5 5-2.24 5-5-2.24-5-5-5zM2 13h2c.55 0 1-.45 1-1s-.45-1-1-1H2c-.55 0-1 .45-1 1s.45 1 1 1zm18 0h2c.55 0 1-.45 1-1s-.45-1-1-1h-2c-.55 0-1 .45-1 1s.45 1 1 1zM11 2v2c0 .55.45 1 1 1s1-.45 1-1V2c0-.55-.45-1-1-1s-1 .45-1 1zm0 18v2c0 .55.45 1 1 1s1-.45 1-1v-2c0-.55-.45-1-1-1s-1 .45-1 1zM5.99 4.58c-.39-.39-1.03-.39-1.41 0-.39.39-.39 1.03 0 1.41l1.06 1.06c.39.39 1.03.39 1.41 0s.39-1.03 0-1.41L5.99 4.58zm12.37 12.37c-.39-.39-1.03-.39-1.41 0-.39.39-.39 1.03 0 1.41l1.06 1.06c.39.39 1.03.39 1.41 0 .39-.39.39-1.03 0-1.41l-1.06-1.06zm1.06-10.96c.39-.39.39-1.03 0-1.41-.39-.39-1.03-.39-1.41 0l-1.06 1.06c-.39.39-.39 1.03 0 1.41s1.03.39 1.41 0l1.06-1.06zM7.05 18.36c.39-.39.39-1.03 0-1.41-.39-.39-1.03-.39-1.41 0l-1.06 1.06c-.39.39-.39 1.03 0 1.41s1.03.39 1.41 0l1.06-1.06z"></path>
-        <path class="system system--dark" d="M9.5 2c-1.82 0-3.53.5-5 1.35 2.99 1.73 5 4.95 5 8.65s-2.01 6.92-5 8.65c1.47.85 3.18 1.35 5 1.35 5.52 0 10-4.48 10-10S15.02 2 9.5 2z"></path>
-      </svg>
+      <slot name="icon-system" part="icon icon-system">
+        <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" width="24" height="24" fill="currentColor" class="default-icon">
+          <path d="M20 3H4c-1.1 0-2 .9-2 2v11c0 1.1.9 2 2 2h3l-1 1v2h12v-2l-1-1h3c1.1 0 2-.9 2-2V5c0-1.1-.9-2-2-2zm0 13H4V5h16v11z"></path>
+          <path class="system system--light" d="M12 7c-2.76 0-5 2.24-5 5s2.24 5 5 5 5-2.24 5-5-2.24-5-5-5zM2 13h2c.55 0 1-.45 1-1s-.45-1-1-1H2c-.55 0-1 .45-1 1s.45 1 1 1zm18 0h2c.55 0 1-.45 1-1s-.45-1-1-1h-2c-.55 0-1 .45-1 1s.45 1 1 1zM11 2v2c0 .55.45 1 1 1s1-.45 1-1V2c0-.55-.45-1-1-1s-1 .45-1 1zm0 18v2c0 .55.45 1 1 1s1-.45 1-1v-2c0-.55-.45-1-1-1s-1 .45-1 1zM5.99 4.58c-.39-.39-1.03-.39-1.41 0-.39.39-.39 1.03 0 1.41l1.06 1.06c.39.39 1.03.39 1.41 0s.39-1.03 0-1.41L5.99 4.58zm12.37 12.37c-.39-.39-1.03-.39-1.41 0-.39.39-.39 1.03 0 1.41l1.06 1.06c.39.39 1.03.39 1.41 0 .39-.39.39-1.03 0-1.41l-1.06-1.06zm1.06-10.96c.39-.39.39-1.03 0-1.41-.39-.39-1.03-.39-1.41 0l-1.06 1.06c-.39.39-.39 1.03 0 1.41s1.03.39 1.41 0l1.06-1.06zM7.05 18.36c.39-.39.39-1.03 0-1.41-.39-.39-1.03-.39-1.41 0l-1.06 1.06c-.39.39-.39 1.03 0 1.41s1.03.39 1.41 0l1.06-1.06z"></path>
+          <path class="system system--dark" d="M9.5 2c-1.82 0-3.53.5-5 1.35 2.99 1.73 5 4.95 5 8.65s-2.01 6.92-5 8.65c1.47.85 3.18 1.35 5 1.35 5.52 0 10-4.48 10-10S15.02 2 9.5 2z"></path>
+        </svg>
+      </slot>
+      <slot name="label-system" part="label label-system">System theme</slot>
     </slot>
   </button>
 `;
 
 class ThemeToggle extends HTMLElement {
-  /**
-   * @private
-   * @type {string[]}
-   * @description List of supported theme states.
-   */
+  /** @type {('light' | 'dark' | 'system')[]} */
   #states = ['light', 'dark', 'system'];
 
-  /**
-   * @private
-   * @type {string}
-   * @description Current theme state.
-   */
+  /** @type {string} */
   #theme;
 
-  /**
-   * @private
-   * @type {number}
-   * @description Current theme state index.
-   */
-  #index;
+  /** @type {number} */
+  #selectedThemeIndex;
 
   constructor() {
     super();
@@ -94,24 +93,6 @@ class ThemeToggle extends HTMLElement {
       this.attachShadow({ mode: 'open' });
       this.shadowRoot.appendChild(template.content.cloneNode(true));
     }
-  }
-
-  static get observedAttributes() {
-    return ['toggle-title'];
-  }
-
-  attributeChangedCallback(name, oldValue, newValue) {
-    if (name === 'toggle-title' && oldValue !== newValue) {
-      this.shadowRoot.getElementById('theme-toggle').setAttribute('title', this.toggleTitle);
-    }
-  }
-
-  get toggleTitle() {
-    return this.getAttribute('toggle-title');
-  }
-
-  set toggleTitle(value) {
-    this.setAttribute('toggle-title', value);
   }
 
   get noStorage() {
@@ -135,12 +116,11 @@ class ThemeToggle extends HTMLElement {
   }
 
   connectedCallback() {
-    this.#upgradeProperty('toggleTitle');
     this.#upgradeProperty('noStorage');
     this.#upgradeProperty('storageKey');
 
     this.#theme = this.#getThemePreference();
-    this.#index = this.#states.indexOf(this.#theme) || 0;
+    this.#selectedThemeIndex = this.#states.indexOf(this.#theme) || 0;
     this.#reflectThemePreference();
 
     this.shadowRoot.getElementById('theme-toggle').addEventListener('click', this.#onClick);
@@ -193,13 +173,12 @@ class ThemeToggle extends HTMLElement {
    */
   #reflectThemePreference() {
     const toggleBtn = this.shadowRoot.getElementById('theme-toggle');
-    toggleBtn.setAttribute('aria-label', this.#theme);
-    toggleBtn.querySelectorAll('slot').forEach(el => el.classList.toggle('hidden', !el.name.includes(this.#theme)));
-    document.documentElement.setAttribute(ROOT_ATTR, this.#theme);
 
-    if (!this.toggleTitle) {
-      toggleBtn.setAttribute('title', `Set theme to ${this.#states[(this.#index + 1) % this.#states.length]}`);
-    }
+    toggleBtn.querySelectorAll(`slot[name="light"], slot[name="dark"], slot[name="system"]`).forEach(el => {
+      el.classList.toggle('hidden', !el.name.startsWith(this.#theme));
+    });
+
+    document.documentElement.setAttribute(ROOT_ATTR, this.#theme);
 
     if (this.#theme === 'system') {
       const prefersDark = window.matchMedia('(prefers-color-scheme: dark)').matches;
@@ -214,8 +193,8 @@ class ThemeToggle extends HTMLElement {
    * @returns {string} Next theme state.
    */
   #nextTheme() {
-    this.#index = (this.#index + 1) % this.#states.length;
-    return this.#states[this.#index];
+    this.#selectedThemeIndex = (this.#selectedThemeIndex + 1) % this.#states.length;
+    return this.#states[this.#selectedThemeIndex];
   }
 
   /**
@@ -227,7 +206,7 @@ class ThemeToggle extends HTMLElement {
     this.#setThemePreference();
     this.#reflectThemePreference();
 
-    this.dispatchEvent(new CustomEvent('theme-toggle:change', {
+    this.dispatchEvent(new CustomEvent('theme-change', {
       bubbles: true,
       composed: true,
       detail: {